Move UpdateChecker to util package
